# Cupy memory hook