--- a/cupy/cuda/__init__.py
+++ b/cupy/cuda/__init__.py
@@ -4,6 +4,7 @@ from cupy.cuda import compiler  # NOQA
 from cupy.cuda import device  # NOQA
 from cupy.cuda import function  # NOQA
 from cupy.cuda import memory  # NOQA
+from cupy.cuda import memory_hook  # NOQA
 from cupy.cuda import pinned_memory  # NOQA
 from cupy.cuda import profiler  # NOQA
 from cupy.cuda import runtime  # NOQA
@@ -57,6 +58,7 @@ from cupy.cuda.memory import Memory  # NOQA
 from cupy.cuda.memory import MemoryPointer  # NOQA
 from cupy.cuda.memory import MemoryPool  # NOQA
 from cupy.cuda.memory import set_allocator  # NOQA
+from cupy.cuda.memory_hook import MemoryHook  # NOQA
 from cupy.cuda.pinned_memory import alloc_pinned_memory  # NOQA
 from cupy.cuda.pinned_memory import PinnedMemory  # NOQA
 from cupy.cuda.pinned_memory import PinnedMemoryPointer  # NOQA