--- a/cupy/cuda/memory.pxd
+++ b/cupy/cuda/memory.pxd
@@ -56,15 +56,18 @@ cdef class PooledMemory(Memory):
 cdef class SingleDeviceMemoryPool:
 
     cdef:
-        object _alloc
+        object _allocator
         dict _in_use
         object _free
         object __weakref__
         object _weakref
         readonly Py_ssize_t _allocation_unit_size
         readonly Py_ssize_t _initial_bins_size
+        readonly int _device_id
 
     cpdef MemoryPointer malloc(self, Py_ssize_t size)
+    cpdef MemoryPointer _alloc(self, Py_ssize_t size)
+    cpdef MemoryPointer _malloc(self, Py_ssize_t size)
     cpdef free(self, size_t ptr, Py_ssize_t size)
     cpdef free_all_blocks(self)
     cpdef free_all_free(self)