--- a/cupy/cuda/memory.pyx
+++ b/cupy/cuda/memory.pyx
@@ -388,7 +388,7 @@ cdef class SingleDeviceMemoryPool:
         self._allocation_unit_size = 512
         self._initial_bins_size = 1024
         self._in_use = {}
-        self._free = [[] for i in range(self._initial_bins_size)]
+        self._free_bins = [[] for i in range(self._initial_bins_size)]
         self._allocator = allocator
         self._weakref = weakref.ref(self)
         self._device_id = device.get_device_id()
@@ -405,12 +405,12 @@ cdef class SingleDeviceMemoryPool:
 
     cpdef void _grow_free_if_necessary(self, Py_ssize_t size):
         """Extend bins (_free) size if necessary"""
-        current_size = len(self._free)
+        current_size = len(self._free_bins)
         if current_size >= size:
             return
         growth_size = size - current_size
         growth = [[] for i in range(growth_size)]
-        self._free.extend(growth)
+        self._free_bins.extend(growth)
 
     cpdef tuple _split(self, Chunk chunk, Py_ssize_t size):
         """Split contiguous block of a larger allocation"""
@@ -432,7 +432,7 @@ cdef class SingleDeviceMemoryPool:
         head.next = remaining
         remaining.prev = head
         index = self._bin_index_from_size(remaining.size)
-        self._free[index].append(remaining)
+        self._free_bins[index].append(remaining)
         return (head, remaining)
 
     cpdef Chunk _merge(self, Chunk head, Chunk remaining):
@@ -458,10 +458,12 @@ cdef class SingleDeviceMemoryPool:
             for hook in hooks_values:
                 hook.alloc_preprocess(device_id, rounded_size)
             try:
+                memptr = None
                 memptr = self._allocator(rounded_size)
             finally:
                 for hook in hooks_values:
-                    hook.alloc_postprocess(device_id, rounded_size)
+                    mem_ptr = memptr.ptr if memptr is not None else 0
+                    hook.alloc_postprocess(device_id, rounded_size, mem_ptr)
             return memptr
         else:
             return self._allocator(rounded_size)
@@ -475,13 +477,30 @@ cdef class SingleDeviceMemoryPool:
             for hook in hooks_values:
                 hook.malloc_preprocess(device_id, size, rounded_size)
             try:
+                memptr = None
                 memptr = self._malloc(rounded_size)
             finally:
                 for hook in hooks_values:
-                    hook.malloc_postprocess(device_id, size, rounded_size)
+                    mem_ptr = memptr.ptr if memptr is not None else 0
+                    hook.malloc_postprocess(device_id, size, rounded_size, mem_ptr)
             return memptr
         else:
             return self._malloc(rounded_size)
+
+    cpdef free(self, size_t ptr, Py_ssize_t size):
+        hooks = memory_hook.get_memory_hooks()
+        if hooks:
+            device_id = self._device_id
+            hooks_values = hooks.values()  # avoid six for performance
+            for hook in hooks_values:
+                hook.free_preprocess(device_id, ptr, size)
+            try:
+                self._free(ptr, size)
+            finally:
+                for hook in hooks_values:
+                    hook.free_postprocess(device_id, ptr, size)
+        else:
+            return self._free(ptr, size)
 
     cpdef MemoryPointer _malloc(self, Py_ssize_t size):
         cdef list free_list = None
@@ -494,9 +513,9 @@ cdef class SingleDeviceMemoryPool:
 
         index = self._bin_index_from_size(size)
         # find best-fit, or a smallest larger allocation
-        length = len(self._free)
+        length = len(self._free_bins)
         for i in range(index, length):
-            free_list = self._free[i]
+            free_list = self._free_bins[i]
             if free_list:
                 chunk = free_list.pop()
                 chunk, _remaining = self._split(chunk, size)
@@ -531,7 +550,7 @@ cdef class SingleDeviceMemoryPool:
         pmem = PooledMemory(chunk, self._weakref)
         return MemoryPointer(pmem, 0)
 
-    cpdef free(self, size_t ptr, Py_ssize_t size):
+    cpdef _free(self, size_t ptr, Py_ssize_t size):
         cdef Chunk chunk
         cdef int index
 
@@ -542,23 +561,23 @@ cdef class SingleDeviceMemoryPool:
         chunk.in_use = False
         if chunk.next and not chunk.next.in_use:
             index = self._bin_index_from_size(chunk.next.size)
-            self._free[index].remove(chunk.next)
+            self._free_bins[index].remove(chunk.next)
             chunk = self._merge(chunk, chunk.next)
 
         if chunk.prev and not chunk.prev.in_use:
             index = self._bin_index_from_size(chunk.prev.size)
-            self._free[index].remove(chunk.prev)
+            self._free_bins[index].remove(chunk.prev)
             chunk = self._merge(chunk.prev, chunk)
 
         index = self._bin_index_from_size(chunk.size)
         self._grow_free_if_necessary(index + 1)
-        self._free[index].append(chunk)
+        self._free_bins[index].append(chunk)
 
     cpdef free_all_blocks(self):
         # Free all **non-split** chunks
         cdef list free_list
         cdef Chunk chunk
-        for free_list in self._free:
+        for free_list in self._free_bins:
             for chunk in free_list:
                 if not chunk.prev and not chunk.next:
                     free_list.remove(chunk)
@@ -571,7 +590,7 @@ cdef class SingleDeviceMemoryPool:
 
     cpdef n_free_blocks(self):
         cdef Py_ssize_t n = 0
-        for v in self._free:
+        for v in self._free_bins:
             n += len(v)
         return n
 
@@ -583,7 +602,7 @@ cdef class SingleDeviceMemoryPool:
 
     cpdef free_bytes(self):
         cdef Py_ssize_t size = 0
-        for free_list in self._free:
+        for free_list in self._free_bins:
             for chunk in free_list:
                 size += chunk.size
         return size