--- a/cupy/cuda/memory_hooks/__init__.py
+++ b/cupy/cuda/memory_hooks/__init__.py
@@ -1,0 +1,4 @@
+from cupy.cuda.memory_hooks import debug_print  # NOQA
+
+# import class and function
+from cupy.cuda.memory_hooks.debug_print import DebugPrintHook  # NOQA