… cupy/cuda/memory_hooks/debug_print.py
… docs/source/reference/cuda.rst
… tests/cupy_tests/cuda_tests/memory_hooks_tests/test_debug_print.py
… tests/cupy_tests/cuda_tests/test_memory_hook.py