for memory profiler.

Code example:


```python
import cupy
from cupy.cuda import memory_hooks

cupy.cuda.set_allocator(cupy.cuda.MemoryPool().malloc)
with memory_hooks.DebugPrintHook():
    x = cupy.array([1, 2, 3])
    del x
```

EDIT: I benchmarked for the case of that no hook is registered as follows:

```
    def test_speed(self):
        started = time.time()
        for i in range(1000000):
            p1 = self.pool.malloc(1)
        elapsed = time.time() - started
        print('elapsed', elapsed)
```

Before: 1.25357346535 sec in average
After: 1.25931959152 sec in average

So, 0.000000006 sec (0.006us) overhead on each malloc operation. 0.4% slowness.
I think this is trivial.