@sonots 2017-07-10
Can anyone review this? I want to get opinions for this design itself first.