@delta2323 2017-07-20
Is it possible to write test code for this functionality?