@sonots 2017-07-20
Added tests for memory hookj