@sonots 2017-07-28
Added `free` hook point and `DebugPrintHook` as a first example of memory hook.
`DebugPrintHook` should be useful for situations like https://github.com/cupy/cupy/issues/317.