@delta2323 2017-08-02
Could you fix autopep8 violation?