@sonots 2017-08-10
> Could you tell me why you make all arguments of XXX_preprocess and XXX_postprocess keyword arguments?

The reason is because it allows us add new parameters in future without breaking lower version compatibility.

https://github.com/cupy/cupy/pull/264/commits/72bebfbc566207c1db6f821c7f8e22e88b35024a