@delta2323 2017-08-10
Could you register `get_memory_hooks`, `MemoryHook`, and `DebugPringHook`.