@delta2323 2017-08-10
> The reason is because it allows us adding new parameters in future without breaking lower version compatibility.
OK, understood.