@sonots 2017-08-10
> Could you register get_memory_hooks, MemoryHook, and DebugPringHook to the docstring?

I registered to cuda.rst except `get_memory_hooks`.
I assume that `get_memory_hooks` is only used by `memory.pyx`.