@sonots 2017-08-11
rebased to master