@delta2323 2017-08-21
@sonots LGTM, I will merge this PR after you resolve the conflict.